perf(v4): split sealed-plan client rendering into client-plans
